Fix: updateId Long to Int conversion in getUpdates

+++ app/src/main/java/com/guardeye/TelegramBot.kt
@@ -104,7 +104,7 @@ object TelegramBot {
                     val text = msg.optString("text", "")
                     val chatObj = msg.optJSONObject("chat")
                     val chatIdFrom = chatObj?.optLong("id", 0L)?.toString() ?: ""
-                    updates.add(Update(text, chatIdFrom, updateId))
+                    updates.add(Update(text, chatIdFrom, updateId.toInt()))
                 }
                 updates
             }
